requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
SQLAlchemy>=2.0.0
psycopg[binary]>=3.1
python-dotenv>=1.0.1
requests>=2.31.0
beautifulsoup4>=4.12.3
gunicorn>=21.2.0
playwright>=1.45.0
textstat>=0.7.3
lxml>=5.1.0
PyJWT>=2.8.0
bcrypt>=4.1.0
PyMySQL>=1.1.0
